requirements.txt
faiss-cpu
openai
numpy
fastapi
uvicorn
